Spy payment service
Spying we use the real object, but still we could override some method calls if we want.

Usually, when testing, we would like to get rid of external points of failure, using mocks or stubs. Seldom do we need to use spying.

In this case, the exceptions are raised in the external calls- should we be testing them in BookingService?